Tweak a failing custom test
* test/lisp/custom-tests.el (custom-tests-require-theme): Fix a
native-comp-related test failure.

In this test, we set load-path to nil, and under native-comp some of
the tests want to give a warning:

Warning (comp): Cannot look-up eln file as no source file was
found for /tmp/custom-tests-bfdbkW/custom-tests--d.elc

But we can't load warning.el after setting load-path to nil, so
require it first.