Fix "Finder" description for multi-file packages
* lisp/finder.el (finder-compile-keywords):  Now that 'version' is
set for more files, it is no longer necessary to check that we have
a version when picking the file from which to take the description.
Doing so caused built-in packages (where the version is known) to have
their package description taken from the last .el file that matched the
file name pattern, usually the last source file in the sub-directory.

This bug was introduced in July 2019 and first released in Emacs 27.1.

Thanks to Michael Albinus for reviewing this patch.