; * lisp/so-long.el: Documentation
Clarify things which are only relevant to 'global-so-long-mode', and
not when the 'so-long' command, or one of its actions, is invoked
directly.

Use the new "customization group" hyperlink syntax.

Only quote 'so-long' as a symbol when it specifically refers to the
command (or will otherwise generate a hyperlink which is appropriate
to the context).  When referring to the library more generally, leave
it unquoted.

Changes from commit 63f419f1339cbd0a7d1e64586854a4f01b3f80d1 are
reverted or improved.