allow party crashers to respond to ical events
If an ical invitation is sent to a mailing list, the recipients
are probably not listed as attendees.  However there are
legitimate situations where these unlisted (or indirectly listed)
recipients are still expected to respond.  RFC5546 allows that,
calling those respondents "party crashers".

* lisp/gnus/gnus-icalendar.el
(gnus-icalendar-event:inline-reply-buttons): Display the response
buttons even if the user was not found in the list of attendees,
but change the labels of those buttons to make clear they are
not explicitly invited.
(gnus-icalendar-event--build-reply-event-body): Add an attendee
line for the user in case one was not found.